edit thuoc tinh voucher

+++ src/main/java/com/example/booker/entity/SanPham.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Entity
@@ -30,7 +31,7 @@ public class SanPham {
     String tac_gia;
     LocalDate ngay_xuat_ban;
     int so_trang;
-    int ma_isbn;
+    BigInteger ma_isbn;
     String phien_ban;
     String anh_san_pham;
     int ma_cua_hang;
